# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Medarx Application API (Phase 1 Privacy Kernel)
  version: 0.1.0-phase1
  summary: HTTP surface for the Medarx Phase 1 privacy kernel.
  description: |
    Medarx is intended for research and demonstration. It performs administrative report assistance using clinician-supplied findings and authorized report text. It does not independently interpret medical images, make diagnostic or treatment recommendations, or autonomously submit clinical reports. Its regulatory classification has not been established, and it is not validated for clinical use.

    This document specifies component J (Application API) of the approved Phase 1 design: the HTTP surface that exercises components A through I. It is the privacy kernel's boundary, not a clinical product surface.

    ## Relationship to the design document

    The normative source for enforcement behaviour is §6 (Fail-Closed Enforcement) of the design. This specification is a faithful encoding of that table; the `layer` values used in the `BlockReceipt` correspond one-to-one with the design's layer designations (`A`, `D.1`, `D.2`, `D.3`, `E`, `F`).

    ## Two things this API deliberately does NOT do

...

...

    ## Function scope

...

    ## Status-code contract

    Two refusal conditions must be distinguishable **by status code alone**, without parsing the body:

    | Status | Meaning | Raised by |
    |---|---|---|
...

...

    ## Request identity

...

    ## Authentication

...

    ## Audit readback

...
  contact:
...
servers:
...
security: []
paths:
...
  /v1/audit/records:
    get:
      tags: [audit]
      operationId: listAuditRecords
      summary: Read back audit records (append-only, raw-PHI-free).
      description: |
        Read-only query surface over component G's append-only, tamper-evident log.

        The response returns **exactly the fields the storage policy persists**: request ID and
        timestamp, function and selected model, de-identification policy version, input and
        approved-payload hashes, redacted field names and action codes, human approval, and final
        disposition.

        It deliberately exposes **no raw-PHI field**, because the audit log never stores one. There is
        no such property on `AuditRecord` to omit, redact, or accidentally leak. Full model prompt
        and response belong to a separate optional restricted, encrypted diagnostic store, which is
        not reachable through this API and is out of Phase 1 scope.

        Records are returned in ascending timestamp order. `chain_verified` reports the result of
        hash-chain verification over the records returned: a relational database alone does not make
        a log immutable, so tampering evidence is verified on read rather than asserted.

        **This readback is not study-scoped, and no filter can make it so in Phase 1.** A caller must
        present a scope to read anything, but the scope is not applied to the page: `AuditRecord` is
        closed and carries no study or patient property, and the audit log's storage policy admits
        none — adding one would put a subject reference into the one store the design requires never
        to hold one. So the effective access control is "present a scope", and a caller holding one
        study's scope reads the whole log. Study-scoped readback is deferred until the storage
        policy can carry a keyed reference. A query parameter this operation does not declare is
        refused with `400` rather than ignored, so a caller cannot believe a filter applied when it
        did not.
      parameters:
        - $ref: '#/components/parameters/RequestIdQueryParam'
        - $ref: '#/components/parameters/FunctionQueryParam'
        - $ref: '#/components/parameters/DispositionQueryParam'
        - $ref: '#/components/parameters/SinceQueryParam'
        - $ref: '#/components/parameters/LimitQueryParam'
      responses:
        '200':
          headers:
            $ref: '#/components/headers/RequestIdHeader'
          description: Matching audit records, oldest first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditRecordCollection'
        '400':
          headers:
            $ref: '#/components/headers/RequestIdHeader'
          $ref: '#/components/responses/MalformedRequest'
        '403':
          headers:
            $ref: '#/components/headers/RequestIdHeader'
          $ref: '#/components/responses/Forbidden'
        '500':
          headers:
            $ref: '#/components/headers/RequestIdHeader'
          $ref: '#/components/responses/UnexpectedError'
  /v1/audit/records/{request_id}:
    get:
      tags: [audit]
      operationId: getAuditRecord
      summary: Read back the audit record for a single request ID.
      description: |
        Returns the append-only audit record for one request ID, with the same stored fields and the
        same absence of raw PHI as the collection endpoint. A request ID that produced both
        processing events and a final disposition returns the merged record.

        If no record exists for the supplied request ID, `404` is returned. The audit log is
        append-only and is never rewritten, so a record that exists cannot disappear.
      parameters:
        - $ref: '#/components/parameters/RequestIdPathParam'
      responses:
        '200':
          headers:
            $ref: '#/components/headers/RequestIdHeader'
          description: The audit record for the supplied request ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditRecord'
        '400':
          headers:
            $ref: '#/components/headers/RequestIdHeader'
          $ref: '#/components/responses/MalformedRequest'
        '403':
          headers:
            $ref: '#/components/headers/RequestIdHeader'
          $ref: '#/components/responses/Forbidden'
        '404':
          headers:
            $ref: '#/components/headers/RequestIdHeader'
          $ref: '#/components/responses/AuditRecordNotFound'
        '500':
          headers:
            $ref: '#/components/headers/RequestIdHeader'
          $ref: '#/components/responses/UnexpectedError'
components:
  securitySchemes: {}
  parameters:
...
    RequestIdPathParam:
      name: request_id
      in: path
      required: true
      description: The request ID of the execution, as returned in the execution response or block receipt.
      schema:
        $ref: '#/components/schemas/RequestId'
...
    RequestIdQueryParam:
      name: request_id
      in: query
      required: false
      description: Filter to a single request ID.
      schema:
        $ref: '#/components/schemas/RequestId'
    FunctionQueryParam:
      name: function
      in: query
      required: false
      description: Filter to one function.
      schema:
        $ref: '#/components/schemas/FunctionName'
    DispositionQueryParam:
      name: disposition
      in: query
      required: false
      description: Filter to one final disposition.
      schema:
        $ref: '#/components/schemas/FinalDisposition'
    SinceQueryParam:
      name: since
      in: query
      required: false
      description: Return only records with a timestamp at or after this instant (RFC 3339).
      schema:
        type: string
        format: date-time
    LimitQueryParam:
      name: limit
      in: query
      required: false
      description: Maximum number of records to return.
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 100
  headers:
    RequestIdHeader:
      description: >-
        The request ID of this exchange, echoed on **every** response including `400`, `403`, `415`
        and `422`. It is the caller's `X-Request-Id` when one was supplied and a server-generated ID
        when one was not, it is the same value the audit record is filed under, and it is present
        even on a response whose body is unparseable — which is the whole reason it is a header and
        not a body field. Declared here rather than left to each response, so the promise is written
        down once.
      schema:
        type: string
  responses:
...
    MalformedRequest:
      headers:
        $ref: '#/components/headers/RequestIdHeader'
      description: >-
        The request could not be understood as a request at all: unparseable body, wrong media type,
        or a JSON document that does not match the request schema. This is a request-construction
        failure (design §6 row 1) and is distinct from both `403` (authorization) and `422`
        (privacy block). Nothing entered the pipeline.

        Where the request is well-formed JSON but violates the schema, the body is a `ValidationError`
        listing the offending fields. Where the body is unparseable, it is a `ProblemDetail`.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            schemaViolation:
              summary: Well-formed JSON that does not match the execution request schema.
              value:
                type: https://medarx.invalid/problems/request-validation
                title: Request does not match the execution request schema
                status: 400
                detail: >-
                  Unrecognized or disallowed properties present: dicom_object, prompt. Medarx does not
                  accept arbitrary DICOM objects or free-form prompts at this surface.
                request_id: 7f3c1a90-2b6e-4f1a-9d55-2c8b7e0a1f34
                errors:
                  - field: body.dicom_object
                    message: Additional properties are not permitted.
                  - field: body.prompt
                    message: Additional properties are not permitted.
            unparseableBody:
              summary: The body could not be parsed as JSON.
              value:
                type: https://medarx.invalid/problems/malformed-request
                title: Malformed request
                status: 400
                detail: Request body is not valid JSON.
                request_id: 7f3c1a90-2b6e-4f1a-9d55-2c8b7e0a1f34
    Forbidden:
      headers:
        $ref: '#/components/headers/RequestIdHeader'
      description: >-
        Authorization failure: the caller is not permitted to execute this function, or the study
        context is outside the caller's authorized scope (design §6 row 1).

        This is `403` and never `422`. It is deliberately distinct from a privacy-policy block: the
        request here was not permitted in the first place, whereas a `422` means the request was
        permitted and well-formed but the privacy pipeline refused it. A caller must be able to tell
        these apart by status code alone, without parsing the body.

        Phase 1 exercises a **synthetic** scope; the design's per-user authorization model (roles,
        per-user permissions) is deferred to Phase 6. No authentication scheme is defined for this
        API.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            functionNotPermitted:
              summary: The function is not permitted for this caller.
              value:
                type: https://medarx.invalid/problems/function-not-permitted
                title: Function not permitted
                status: 403
                detail: The requested function is not permitted in the caller's scope.
                request_id: 7f3c1a90-2b6e-4f1a-9d55-2c8b7e0a1f34
            studyScopeUnauthorized:
              summary: The study context is outside the caller's authorized scope.
              value:
                type: https://medarx.invalid/problems/study-scope-unauthorized
                title: Study scope not authorized
                status: 403
                detail: The requested study is not within the caller's authorized scope.
                request_id: 7f3c1a90-2b6e-4f1a-9d55-2c8b7e0a1f34
...
    AuditRecordNotFound:
      headers:
        $ref: '#/components/headers/RequestIdHeader'
      description: No audit record exists for the supplied request ID.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    UnexpectedError:
      headers:
        $ref: '#/components/headers/RequestIdHeader'
      description: >-
        An unexpected server-side failure. Sensitive-data filtering is applied to application logs,
        exception traces, and tracing spans under the same policy as the audit log, so no raw PHI is
        disclosed through this response or its logs.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
  schemas:
    FunctionName:
      type: string
      title: FunctionName
      description: >-
        The closed set of Medarx functions. The set is closed by design: functions are a fixed,
        enumerated surface, not free-form.

        Phase 1 implements the **privacy pipeline around** these functions, not their full clinical
        behaviour. Phase 1 exercises the pipeline with report text and metadata; `Draft`'s automated
        output validation is a Phase 4 concern, and `Prior Summary` / `Ask` contract validation is a
        Phase 5 concern. What Phase 1 proves for each is authorization, allowlisted extraction,
        pseudonymization, layered redaction, and policy approval or refusal before transmission.
      enum:
        - Draft
        - Prior Summary
        - Ask
    RequestId:
      type: string
      description: >-
        Identifier of a single API request. Generated by the server when the caller does not supply
        one; carried on every block receipt, every successful execution response, and every audit
        record.
      examples: [7f3c1a90-2b6e-4f1a-9d55-2c8b7e0a1f34]
    PolicyVersion:
      type: string
      description: >-
        The de-identification policy version in force. Appears in block receipts, execution
        responses, and audit records. An unknown or missing policy version causes a fail-closed block
        in every implemented mode.

        This is a free string and every value in this document is illustrative: `ps315-baseline-v1`
        and the values in the `Blocked` examples name the *shape* of a receipt, not what a given
        deployment publishes. A deployment publishes the version it is configured with, so a receipt
        read against a running system carries that deployment's own string.
      examples: [ps315-baseline-v1]
    PolicyMode:
      type: string
      description: >-
        The configured policy mode (component E). Fixed per deployment; never selectable per request.
      enum:
        - strict_local
        - cloud
        - authorized_local
...
    ActionCode:
      type: string
      description: >-
        A non-raw disposition code from component G's action-code table. Codes classify what was
        done or refused; they never carry a raw value, so a block receipt is safe to log and return
        in full.

        The set is an open enumeration at the schema level because G's table is the authoritative
        source and is expected to grow. The values below are **not** all emitted in Phase 1, and the
        five that are not fall into two groups that are tracked differently.

        Three are advertised by a published example, so the example-reachability sweep can see that
        they are unreachable: `ARBITRARY_DICOM_OBJECT_REJECTED` and `FREE_FORM_PROMPT_REJECTED`,
        which belong to the application API's request surface (component J, not written yet), and
        `UNRESOLVED_DISPOSITION`, which is declared and deliberately unemitted because component E
        blocks an unresolved disposition under the disposition's own code. Each of the three is
        listed as pending in `backend/tests/test_contract_schemas.py::_NOT_YET_EMITTED`, and that
        set is held equal to the pairs the contract's own examples advertise, so it cannot grow to
        cover a real failure.

        Two are not advertised anywhere: `FUNCTION_NOT_PERMITTED` and `UNAUTHORIZED_SCOPE` are
        reserved for component J, and no published example carries them. A code that appears in no
        example is invisible to the example-reachability sweep by construction — there is no
        example for it to be unreachable *in* — so these two are declared in
        `backend/tests/test_contract_schemas.py::_RESERVED_NOT_ADVERTISED` instead and are checked
        in the one direction that is available to them: that the package has not started emitting
        them. The gap is recorded rather than smoothed: until J is written, these two are reserved
        by this description and by G's code table alone, and nothing would catch a component that
        began emitting one without also declaring it.

        A code published for a block the kernel cannot produce would be a claim about behaviour, so
        this description does not make one.
      examples:
        - NER_UNRESOLVED
      enum:
        - NER_UNRESOLVED
        - DETERMINISTIC_REPLACEMENT_FAILED
        - CONTRACT_VIOLATION
        - HASH_MISMATCH
        - LEFTOVER_PATTERN_MATCH
        - UNSHIFTED_DATE
        - MISSING_SURROGATE
        - FIELD_NOT_ALLOWLISTED
        - UNKNOWN_FUNCTION
        - UNRESOLVED_DISPOSITION
        - UNKNOWN_POLICY_VERSION
        - POLICY_CONFIG_ERROR
        - PAYLOAD_MISMATCH
        - UNKNOWN_MODEL
        - ARBITRARY_DICOM_OBJECT_REJECTED
        - FREE_FORM_PROMPT_REJECTED
        - UNAUTHORIZED_SCOPE
        - FUNCTION_NOT_PERMITTED
        - UNKNOWN_DICOM_ATTRIBUTE
        - UNRESOLVED_EMPTY_BODY
        - UNAPPROVED_PAYLOAD
        - MALFORMED_METADATA
        - SURROGATE_SHAPED_REFERENCE_REJECTED
        - LOW_CONFIDENCE_NER_UNRESOLVED
...
    ProblemDetail:
      type: object
      description: >-
        RFC 9457 problem document, used for `400`, `403`, `404`, `409`, and `415` responses. Used in
        preference to the `BlockReceipt` wherever the request was never a valid request; a privacy
        block always uses the `BlockReceipt` instead.
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: Stable problem-type URI.
        title:
          type: string
          description: Short human-readable summary of the problem type.
        status:
          type: integer
          description: The HTTP status code, repeated for clients that do not read the status line.
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        request_id:
          $ref: '#/components/schemas/RequestId'
        errors:
          type: array
          description: Per-field detail, present on schema violations.
          items:
            $ref: '#/components/schemas/ValidationError'
...
    FinalDisposition:
      type: string
      description: >-
        The end state of a request as recorded in the audit log.
      enum:
        - approved
        - blocked
        - pending_human_approval
        - approved_by_human
        - rejected_by_human
...
    AuditRecord:
      type: object
      additionalProperties: false
      required:
        - request_id
        - timestamp
        - function
        - policy_version
        - final_disposition
      description: >-
        One record from the append-only, tamper-evident audit log.

        The properties below are **exactly** the fields the storage policy persists: request ID and
        timestamp, function and selected model, de-identification policy version, input and
        approved-payload hashes, redacted field names and action codes, human approval, and final
        disposition.

        There is no raw-PHI property on this schema, and that absence is the design, not an omission
        in the documentation: raw PHI is never stored in the normal audit log, so there is no such
        field to return. Full model prompt and response live only in a separate optional restricted,
        encrypted diagnostic store, which this API does not expose.
      properties:
        request_id:
          $ref: '#/components/schemas/RequestId'
        timestamp:
          type: string
          format: date-time
          description: When the event was appended.
        function:
          $ref: '#/components/schemas/FunctionName'
        selected_model:
          type: string
          description: The model used, if the request reached the gateway. Absent for a blocked request.
        policy_version:
          $ref: '#/components/schemas/PolicyVersion'
        policy_mode:
          $ref: '#/components/schemas/PolicyMode'
        input_hash:
          type: string
          description: Hash of the caller's input as received.
        approved_payload_hash:
          type: string
          description: >-
            Hash of the exact approved payload transmitted. Absent for a blocked request, since
            nothing was approved.
        redacted_field_names:
          type: array
          description: >-
            Names of the fields the redaction layers acted on. Field **names** only — never the
            values they held. This is a deliberate record of what was processed without becoming a
            second store of what was in it.
          items:
            type: string
          examples:
            - [report_text, dicom_metadata.patient_id, dicom_metadata.accession_number]
        action_codes:
          type: array
          description: Non-raw disposition codes recorded for this request.
          items:
            $ref: '#/components/schemas/ActionCode'
        human_approval:
          oneOf:
            - $ref: '#/components/schemas/HumanApprovalState'
            - type: 'null'
          description: >-
            The recorded human sign-off, or `null` while the draft is still awaiting review. A
            `null` here means "not yet decided", never "decided not to review".
        final_disposition:
          $ref: '#/components/schemas/FinalDisposition'
        chain_verified:
          type: boolean
          description: >-
            Result of hash-chain verification over the records returned by this query. A relational
            database alone does not make a log immutable, so tamper evidence is verified on read
            rather than asserted.
          examples: [true]
        layer:
          oneOf:
            - $ref: '#/components/schemas/Layer'
            - type: 'null'
          description: >-
            The component that refused the request, or `null` when nothing refused it, which is the
            approved case. A classifier with no value in it, and the only field on this schema from
            which `stages` is derived.
        stages:
          type: array
          readOnly: true
          description: >-
            The pipeline stages this request reached, derived at read time from `layer` and not
            stored: the prefix of `PipelineStage` up to and including the stage owning that layer,
            and all six when there is no layer. A request blocked at `D.2` reached three stages and
            never reached `Policy decision` or `Model request`; deriving it is what keeps the two
            from being able to disagree.
          items:
            $ref: '#/components/schemas/PipelineStage'
...
    AuditRecordCollection:
      type: object
      additionalProperties: false
      required:
        - records
        - count
      description: A page of audit records, oldest first.
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/AuditRecord'
        count:
          type: integer
          description: Number of records in this response.
        chain_verified:
          type: boolean
          description: >-
            Aggregate result of hash-chain verification across the returned records. `false` means
            at least one record in the chain did not verify and the log's integrity is in question.
        input_source:
          type: string
          description: >-
            Attributed input source for the matched records, where the caller recorded one.

            **Always absent in Phase 1.** `AuditRecord` is closed and has no input-source field, and
            the audit log's storage policy admits none, so there is no stored value for this member
            to report; populating it needs a new field on a record whose policy says no field may
            hold free text. It is left declared because it is a documented part of this collection,
            and deleting it would be a change with nothing behind it.
          enum:
            - synthetic_corpus
            - clinician_supplied